requirements.txt
pystray>=0.19.5
Pillow>=10.0.0
pywin32>=306
plyer>=2.1.0
